+++ packages/bridge-controller/src/bridge-controller.ts
@@ -35,6 +35,7 @@ import type {
   BridgeControllerState,
   BridgeControllerMessenger,
   FetchFunction,
+  InputPrimaryDenomination,
 } from './types';
 import { getAssetIdsForToken, toExchangeRates } from './utils/assets';
 import { hasSufficientBalance } from './utils/balance';
@@ -156,6 +157,12 @@ const metadata: StateMetadata<BridgeControllerState> = {
     includeInDebugSnapshot: false,
     usedInUi: true,
   },
+  inputPrimaryDenomination: {
+    includeInStateLogs: true,
+    persist: true,
+    includeInDebugSnapshot: false,
+    usedInUi: true,
+  },
   quoteStreamComplete: {
     includeInStateLogs: true,
     persist: false,
@@ -195,6 +202,7 @@ const MESSENGER_EXPOSED_METHODS = [
   'updateBatchSellTrades',
   'stopPollingForQuotes',
   'setLocation',
+  'setInputPrimaryDenomination',
   'resetState',
   'setChainIntervalLength',
   'trackUnifiedSwapBridgeEvent',
@@ -711,6 +719,14 @@ export class BridgeController extends StaticIntervalPollingController<BridgePoll
     this.#location = location;
   };
 
+  setInputPrimaryDenomination = (
+    inputPrimaryDenomination: InputPrimaryDenomination,
+  ) => {
+    this.update((state) => {
+      state.inputPrimaryDenomination = inputPrimaryDenomination;
+    });
+  };
+
   resetState = (
     reason = AbortReason.ResetState,
     quoteRequestIndex: number | null = null,
@@ -1155,6 +1171,9 @@ export class BridgeController extends StaticIntervalPollingController<BridgePoll
       location: clientProps?.location ?? this.#location,
       action_type: MetricsActionType.SWAPBRIDGE_V1,
     };
+    const inputPrimaryDenominationProperties = {
+      input_primary_denomination: this.state.inputPrimaryDenomination,
+    };
     const quoteRequest = this.state.quoteRequest[quoteRequestIndex];
     switch (eventName) {
       case UnifiedSwapBridgeEventName.ButtonClicked:
@@ -1172,6 +1191,16 @@ export class BridgeController extends StaticIntervalPollingController<BridgePoll
             this.state.tokenSecurityTypeDestination,
           ),
           ...this.#getRequestMetadata(),
+          ...inputPrimaryDenominationProperties,
+          ...baseProperties,
+        };
+      case UnifiedSwapBridgeEventName.FiatCryptoToggleClicked:
+        return {
+          ...getRequestParams(
+            quoteRequest,
+            this.state.tokenSecurityTypeDestination,
+          ),
+          swap_type: getSwapTypeFromQuote(quoteRequest),
           ...baseProperties,
         };
       case UnifiedSwapBridgeEventName.QuotesValidationFailed:
@@ -1192,6 +1221,7 @@ export class BridgeController extends StaticIntervalPollingController<BridgePoll
           ...this.#getRequestMetadata(),
           ...this.#getQuoteFetchData(),
           refresh_count: this.state.quotesRefreshCount,
+          ...inputPrimaryDenominationProperties,
           ...baseProperties,
         };
       case UnifiedSwapBridgeEventName.QuotesRequested:
@@ -1202,6 +1232,7 @@ export class BridgeController extends StaticIntervalPollingController<BridgePoll
           ),
           ...this.#getRequestMetadata(),
           has_sufficient_funds: !quoteRequest.insufficientBal,
+          ...inputPrimaryDenominationProperties,
           ...baseProperties,
         };
       case UnifiedSwapBridgeEventName.QuotesError: